requirements.txt
pdfplumber
numpy
scikit-learn
groq
langgraph
langchain-core
chromadb
sentence-transformers
python-dotenv
streamlit

